+++ tests/repository/metadata/meta_common.py
@@ -755,3 +755,21 @@ def test_delete_model(meta: MetadataRepository, project: Project, task: Task, mo
 def test_delete_not_existing_model(meta: MetadataRepository, model: Model):
     with pytest.raises(NonExistingModelError):
         meta.delete_model(model)
+
+
+def test_inner_objects_binded(meta: MetadataRepository, project: Project, task: Task, model: Model):
+    task.project = meta.create_project(project)
+    meta.create_task(task)
+
+    model.task_id = task.id
+    model = meta.create_model(model)
+
+    new_project = meta.get_project_by_name(project.name)
+    assert new_project.id is not None
+    assert new_project.has_meta_repo
+    new_task = new_project.tasks(task.name)
+    assert new_task.id is not None
+    assert new_task.has_meta_repo
+    new_model = new_task.models(model.name)
+    assert new_model.id is not None
+    assert new_model.has_meta_repo